pyproject.toml
[project]
name = "faq-chatbot"
version = "0.1.0"
description = "RAG FAQ chatbot"
readme = "README.md"
requires-python = ">=3.11,<3.14"

dependencies = [
    "langchain>=1.2.0",
    "langchain-core>=1.2.0",
    "langchain-community>=0.3.0",
    "pydantic>=2.7,<3.0",
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "ollama>=0.3.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "streamlit>=1.56.0",
]
